Fix for arguments with value as None not passed by decorators
Fix for arguments with value as None not passed by decorators.
When an argument value is None, decorators are treating it as missing argument. So removing the check for None.

NOTE:
When the arguments are actually missing, this change will pass them also as None, which may be an undesired behavior.